✨ feature : 기록 초기화
세부 삭제 기능 구현 필요

> Related to : #9

+++ src/main/java/great/job/mytamin/domain/user/service/UserService.java
@@ -3,6 +3,7 @@ package great.job.mytamin.domain.user.service;
 import great.job.mytamin.domain.myday.service.DaynoteService;
 import great.job.mytamin.domain.myday.service.WishService;
 import great.job.mytamin.domain.mytamin.service.MytaminService;
+import great.job.mytamin.domain.user.dto.request.InitRequest;
 import great.job.mytamin.domain.user.dto.request.ProfileUpdateRequest;
 import great.job.mytamin.domain.user.dto.response.ProfileResponse;
 import great.job.mytamin.domain.user.entity.User;
@@ -65,12 +66,25 @@ public class UserService {
     기록 초기화
     */
     @Transactional
-    public void deleteAll(User user) {
-        user.initData(); // 숨 고르기, 감각 깨우기 데이터 초기화
-        mytaminService.deleteAll(user); // 마이타민 (칭찬 처방, 하루 진단) 삭제
-        daynoteService.deleteAll(user); // 데이노트 삭제
-        wishService.deleteAll(user); // 위시 삭제
-        userRepository.save(user);
+    public void initData(User user, InitRequest initRequest) {
+        //mytaminService.deleteAll(user); // 마이타민 (칭찬 처방, 하루 진단) 삭제
+        
+        if(initRequest.isInitReport()) {
+            user.initData(); // 숨 고르기, 감각 깨우기 데이터 초기화
+            userRepository.save(user);
+            
+            // Report 초기화 로직
+        }
+        if(initRequest.isInitCare()) {
+            user.initData(); // 숨 고르기, 감각 깨우기 데이터 초기화
+            userRepository.save(user);
+            
+            // Care 초기화 로직
+        }
+        if(initRequest.isInitMyday()) {
+            daynoteService.deleteAll(user);
+            wishService.deleteAll(user);
+        }
     }
 
     /*
@@ -78,7 +92,7 @@ public class UserService {
     */
     @Transactional
     public void withdraw(User user) {
-        deleteAll(user); // 기록 초기화
+        initData(user, new InitRequest(true, true, true)); // 기록 초기화
         awsS3Service.deleteImg(user.getProfileImgUrl()); // 프로필 이미지 삭제
         userRepository.delete(user); // 유저 삭제
     }
